requirements.txt
django
django_q
virtualenv
python-binance